java/jni refactoring: move special types handling to separate module
